requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PlexAPI==4.15.6
python-dotenv==1.0.0
requests==2.31.0
